requirements.txt
Flask==1.0.2
selenium==3.141.0
pytest==5.4.1
pure_python_adb==0.2.3.dev0
nested_lookup==0.1.5
requests==2.22.0
pytest-asyncio==0.12.0
flaky==3.6.1
allure-pytest==2.8.13
pillow==7.1.2
pytesseract==0.3.4
pytest-dependency==0.5.1
pytest-json-report==1.2.1